move titles to a config file
