+++ src/helpers/config/pageTitles.js
@@ -1,0 +1,9 @@
+const titles = {
+  '/latest-rules/': 'Latest Rules',
+  '/user/': 'User Rules',
+  '/orphaned/': 'Orphaned Rules',
+  '/archived/': 'Archived Rules',
+  '/profile/': 'Profile',
+};
+
+export default titles;
